+++ packages/cryptoassets/src/chains/testnet/evm/ethereum.ts
@@ -1,3 +1,4 @@
+import { NftProviderType } from '../../../types';
 import ethereum from '../../../chains/mainnet/evm/ethereum';
 import { transformMainnetToTestnetChain } from '../../../chains/utils';
 
@@ -9,14 +10,15 @@ export default transformMainnetToTestnetChain(
     networkId: 3,
     chainId: 3,
     isTestnet: true,
-    rpcUrls: ['https://ropsten.infura.io/v3/da99ebc8c0964bb8bb757b6f8cc40f1f'],
+    rpcUrls: ['https://ropsten.infura.io/v3/a2ad6f8c0e57453ca4918331f16de87d'],
   },
   [
     {
-      tx: 'https://ropsten.etherscan.io/tx/',
-      address: 'https://ropsten.etherscan.io/address/',
-      token: 'https://ropsten.etherscan.io/token/',
+      tx: 'https://ropsten.etherscan.io/tx/{hash}',
+      address: 'https://ropsten.etherscan.io/address/{address}',
+      token: 'https://ropsten.etherscan.io/token/{token}',
     },
   ],
-  'https://faucet.metamask.io/'
+  'https://faucet.metamask.io/',
+  NftProviderType.Moralis
 );