+++ packages/wallet-core/src/build.config.ts
@@ -31,9 +31,6 @@ export interface WalletCoreConfig {
   batchEsploraApis: {
     [key in Network]: string;
   };
-  rskRpcUrls: {
-    [key in Network]: string;
-  };
   nameResolvers: {
     uns: {
       resolutionService: string;
@@ -95,10 +92,6 @@ const config: WalletCoreConfig = {
   batchEsploraApis: {
     testnet: 'https://electrs-batch-testnet-api.liq-chainhub.net/',
     mainnet: 'https://electrs-batch-mainnet-api.liq-chainhub.net/',
-  },
-  rskRpcUrls: {
-    testnet: 'https://testnet.sovryn.app/rpc',
-    mainnet: 'https://mainnet.sovryn.app/rpc',
   },
   swapProviders: {
     testnet: {
@@ -247,20 +240,7 @@ const config: WalletCoreConfig = {
   },
   discordUrl: 'https://discord.gg/Xsqw7PW8wk',
   networks: [Network.Mainnet, Network.Testnet],
-  chains: [
-    ChainId.Bitcoin,
-    ChainId.Ethereum,
-    ChainId.Rootstock,
-    ChainId.BinanceSmartChain,
-    ChainId.Near,
-    ChainId.Polygon,
-    ChainId.Arbitrum,
-    ChainId.Terra,
-    ChainId.Fuse,
-    ChainId.Avalanche,
-    ChainId.Solana,
-    ChainId.Optimism,
-  ],
+  chains: Object.values(ChainId),
   supportedBridgeAssets: ['RBTC', 'AVAX'],
   nameResolvers: {
     uns: {